soc: qcom: hab: unify variable type to avoid implicit conversion
correcting the type of sizebytes and total_size to size_t,it can avoid
implicit conversion which maybe cause some unknown issue.

+++ drivers/soc/qcom/hab/ghs_comm.c
@@ -39,7 +39,7 @@ int physical_channel_send(struct physical_channel *pchan,
 		struct hab_header *header,
 		void *payload)
 {
-	int sizebytes = HAB_HEADER_GET_SIZE(*header);
+	size_t sizebytes = HAB_HEADER_GET_SIZE(*header);
 	struct ghs_vdev *dev  = (struct ghs_vdev *)pchan->hyp_data;
 	GIPC_Result result;
 	uint8_t *msg;
